DZ 17 lesson

+++ create_data.py
@@ -20,7 +20,7 @@ class Movie(db.Model):
     description = db.Column(db.String(255))
     trailer = db.Column(db.String(255))
     year = db.Column(db.Integer)
-    rating = db.Column(db.Integer)
+    rating = db.Column(db.Float)
     genre_id = db.Column(db.Integer, db.ForeignKey("genre.id"))
     genre = db.relationship("Genre")
     director_id = db.Column(db.Integer, db.ForeignKey("director.id"))
@@ -29,13 +29,17 @@ class Movie(db.Model):
 
 class Director(db.Model):
     __tablename__ = 'director'
+
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
 
+
 class Genre(db.Model):
     __tablename__ = 'genre'
+
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(255))
+
 
 db.drop_all()
 db.create_all()
